python-service/requirements.txt
# Face Recognition Service Dependencies
# Python 3.11+

fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1

# InsightFace and dependencies
insightface==0.7.3
onnxruntime==1.20.1

# Image processing
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4

# HTTP client (for health checks)
httpx==0.27.2

# Logging
python-json-logger==2.0.7
